# Storage Trie optimizations