This PR introduces a few small optimizations in the `StorageTrie` component.

## Changes

- `Cache` that includes the precomputed Keccaks of indexes is now a simple array that allows faster indexing
   - improvement: 40ns on looking up all the values
- getting a `StorageCell` value will memoize the calculated key in a small cache. This should help in addressing scenarios like ERC20 when the storage access provides the Read-Update-Write behavior and the same cell is re-entered to store the value.  For ERC20, like `WrappedEther` for a transfer it will be two pairs of Read-Update-Write.
   - improvement at range of a fraction of a microsecond for a single RUW

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [x] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [x] Yes
- [ ] No

#### If yes, did you write tests?

- [ ] Yes
- [ ] No

#### Notes on testing

Compare timing with a node from `master`

## Documentation

#### Requires documentation update

- [ ] Yes
- [x] No

#### Requires explanation in Release Notes

- [ ] Yes
- [x] No

## Benchmarks

```bash
BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22621.2428/22H2/2022Update/SunValley2)
12th Gen Intel Core i9-12900HK, 1 CPU, 20 logical and 14 physical cores
.NET SDK=7.0.113
  [Host]     : .NET 7.0.13 (7.0.1323.51816), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.13 (7.0.1323.51816), X64 RyuJIT AVX2
```

### Before

|                                    Method |        Mean |     Error |    StdDev |
|------------------------------------------ |------------:|----------:|----------:|
|                        CellIndexes_cached |    942.4 ns |   7.72 ns |   6.84 ns |
|           CellIndexes_memoized_for_writes | 39,835.2 ns | 424.63 ns | 397.20 ns |
| CellIndexes_beyond_memoization_for_writes | 78,823.6 ns | 472.35 ns | 418.73 ns |

### After

|                                    Method |        Mean |     Error |    StdDev |
|------------------------------------------ |------------:|----------:|----------:|
|                        CellIndexes_cached |    907.1 ns |  17.82 ns |  26.68 ns |
|           CellIndexes_memoized_for_writes | 35,375.8 ns | 335.33 ns | 280.01 ns |
| CellIndexes_beyond_memoization_for_writes | 77,059.5 ns | 441.10 ns | 344.38 ns |