@Scooletz 2023-10-30
This cache, to make it smaller and not attached to a tree, could be pushed to `PersistentStorageProvider` as it sets the values into the tree. It would require amending two callsites, one for reading and one for writing:

https://github.com/NethermindEth/nethermind/blob/9e3aa25d63aa9a62595930959ad4cfd24b5a39dd/src/Nethermind/Nethermind.State/PersistentStorageProvider.cs#L241

And

https://github.com/NethermindEth/nethermind/blob/9e3aa25d63aa9a62595930959ad4cfd24b5a39dd/src/Nethermind/Nethermind.State/PersistentStorageProvider.cs#L172

Depending on the size it could work, but also would probably result in many collisions when one read is not followed immiedietaly by the write as the changes are waited to be applied to the tree later. To consider.